spit ConcurrentTestGwt into two parts
